update event handler test case
